+++ castor/lib/core/src/main/scala/io/iohk/atala/castor/core/model/ProtoModelHelper.scala
@@ -18,6 +18,8 @@ import io.iohk.atala.castor.core.model.did.{
   ScheduledDIDOperationStatus,
   Service,
   ServiceType,
+  SignedPrismDIDOperation,
+  UpdateDIDAction,
   VerificationRelationship
 }
 import io.iohk.atala.prism.crypto.EC
@@ -40,18 +42,76 @@ private[castor] trait ProtoModelHelper {
     def toProto: ByteString = ByteString.copyFrom(bytes)
   }
 
-  extension (createDIDOp: PrismDIDOperation.Create) {
+  extension (signedOperation: SignedPrismDIDOperation) {
+    def toProto: node_models.SignedAtalaOperation =
+      node_models.SignedAtalaOperation(
+        signedWith = signedOperation.signedWithKey,
+        signature = signedOperation.signature.toArray.toProto,
+        operation = Some(signedOperation.operation.toAtalaOperation)
+      )
+  }
+
+  extension (operation: PrismDIDOperation.Create) {
     def toProto: node_models.AtalaOperation.Operation.CreateDid = {
       node_models.AtalaOperation.Operation.CreateDid(
         value = node_models.CreateDIDOperation(
           didData = Some(
             node_models.CreateDIDOperation.DIDCreationData(
-              publicKeys = createDIDOp.publicKeys.map(_.toProto) ++ createDIDOp.internalKeys.map(_.toProto),
-              services = createDIDOp.services.map(_.toProto)
+              publicKeys = operation.publicKeys.map(_.toProto) ++ operation.internalKeys.map(_.toProto),
+              services = operation.services.map(_.toProto)
             )
           )
         )
       )
+    }
+  }
+
+  extension (operation: PrismDIDOperation.Deactivate) {
+    def toProto: node_models.AtalaOperation.Operation.DeactivateDid = {
+      node_models.AtalaOperation.Operation.DeactivateDid(
+        value = node_models.DeactivateDIDOperation(
+          previousOperationHash = operation.previousOperationHash.toArray.toProto,
+          id = operation.did.suffix.toString
+        )
+      )
+    }
+  }
+
+  extension (operation: PrismDIDOperation.Update) {
+    def toProto: node_models.AtalaOperation.Operation.UpdateDid = {
+      node_models.AtalaOperation.Operation.UpdateDid(
+        value = node_models.UpdateDIDOperation(
+          previousOperationHash = operation.previousOperationHash.toArray.toProto,
+          id = operation.did.suffix.toString,
+          actions = operation.actions.map(_.toProto)
+        )
+      )
+    }
+  }
+
+  extension (action: UpdateDIDAction) {
+    def toProto: node_models.UpdateDIDAction = {
+      val a = action match {
+        case UpdateDIDAction.AddKey(publicKey) =>
+          node_models.UpdateDIDAction.Action.AddKey(node_models.AddKeyAction(Some(publicKey.toProto)))
+        case UpdateDIDAction.AddInternalKey(publicKey) =>
+          node_models.UpdateDIDAction.Action.AddKey(node_models.AddKeyAction(Some(publicKey.toProto)))
+        case UpdateDIDAction.RemoveKey(id) =>
+          node_models.UpdateDIDAction.Action.RemoveKey(node_models.RemoveKeyAction(id))
+        case UpdateDIDAction.AddService(service) =>
+          node_models.UpdateDIDAction.Action.AddService(node_models.AddServiceAction(Some(service.toProto)))
+        case UpdateDIDAction.RemoveService(id) =>
+          node_models.UpdateDIDAction.Action.RemoveService(node_models.RemoveServiceAction(id))
+        case UpdateDIDAction.UpdateService(serviceId, serviceType, endpoints) =>
+          node_models.UpdateDIDAction.Action.UpdateService(
+            node_models.UpdateServiceAction(
+              serviceId = serviceId,
+              `type` = serviceType.fold("")(_.name),
+              serviceEndpoints = endpoints.map(_.toString)
+            )
+          )
+      }
+      node_models.UpdateDIDAction(action = a)
     }
   }
 
@@ -147,10 +207,10 @@ private[castor] trait ProtoModelHelper {
       Clock.instant.map { now =>
         didData
           .withPublicKeys(didData.publicKeys.filter { publicKey =>
-            publicKey.revokedOn.flatMap(_.toInstant).forall(revokeTime => revokeTime isBefore now)
+            publicKey.revokedOn.flatMap(_.toInstant).forall(revokeTime => revokeTime isAfter now)
           })
           .withServices(didData.services.filter { service =>
-            service.deletedOn.flatMap(_.toInstant).forall(revokeTime => revokeTime isBefore now)
+            service.deletedOn.flatMap(_.toInstant).forall(revokeTime => revokeTime isAfter now)
           })
       }
     }