+++ castor/lib/core/src/main/scala/io/iohk/atala/castor/core/util/Prelude.scala
@@ -1,0 +1,10 @@
+package io.iohk.atala.castor.core.util
+
+// consider moving this to shared library
+object Prelude {
+
+  extension [T](xs: Seq[T]) {
+    def isUnique: Boolean = xs.length == xs.distinct.length
+  }
+
+}