+++ mercury/mercury-library/agent/src/main/scala/io/iohk/atala/mercury/CoordinateMediationPrograms.scala
@@ -7,7 +7,7 @@ import io.circe.syntax._
 import io.circe.parser._
 import io.circe.JsonObject
 
-import io.iohk.atala.mercury.{_, given}
+import io.iohk.atala.mercury.{given, _}
 import io.iohk.atala.mercury.model._
 import io.iohk.atala.mercury.protocol.coordinatemediation._
 import io.iohk.atala.mercury.protocol.invitation.v2.Invitation
@@ -31,17 +31,18 @@ object CoordinateMediationPrograms {
   def senderMediationRequestProgram(mediatorURL: String = "http://localhost:8000") = {
 
     for {
-      _ <- ZIO.log("#### Send Mediation request  ####")
+      _ <- ZIO.log("#### Send Mediation request ####")
       link <- InvitationPrograms
         .getInvitationProgram(mediatorURL + "/oob_url")
         .map(_.toOption) // FIXME
-      agentService <- ZIO.service[DidComm]
+      opsService <- ZIO.service[DidOps]
+      agentService <- ZIO.service[DidAgent]
 
-      planMessage = link.map(to => replyToInvitation(agentService.myDid, to)).get
+      planMessage = link.map(to => replyToInvitation(agentService.id, to)).get
       invitationFrom = link.get.from
       _ <- ZIO.log(s"Invitation from $invitationFrom")
 
-      encryptedMessage <- agentService.packEncrypted(planMessage, to = invitationFrom)
+      encryptedMessage <- opsService.packEncrypted(planMessage, to = invitationFrom)
       _ <- ZIO.log("Sending bytes ...")
       jsonString = encryptedMessage.string
       _ <- ZIO.log(jsonString)
@@ -50,7 +51,7 @@ object CoordinateMediationPrograms {
       res <- client.postDIDComm(mediatorURL, jsonString)
       _ <- ZIO.log(res.bodyAsString)
 
-      messageReceived <- agentService.unpack(res.bodyAsString)
+      messageReceived <- opsService.unpack(res.bodyAsString)
       _ <- Console.printLine("Unpacking and decrypting the received message ...")
       _ <- Console.printLine("*" * 100)
       _ <- Console.printLine(toPrettyJson(messageReceived.getMessage.toString))