+++ mercury/mercury-library/protocol-invitation/src/test/scala/io/iohk/atala/mercury/protocol/invitation/v1/InvitationV1Spec.scala
@@ -16,7 +16,7 @@ class InvitationV1Spec extends ZSuite {
 
   test("out-of-band invitation") {
     val payload = "attachmentData"
-    val payloadBase64 = java.util.Base64.getUrlEncoder.encodeToString(Json.fromString(payload).noSpaces.getBytes)
+    val payloadBase64 = java.util.Base64.getUrlEncoder.encodeToString(payload.getBytes)
 
     val expectedJson = parse(s"""{
                                 |  "@id": "f3375429-b116-4224-b55f-563d7ef461f1",
@@ -35,7 +35,6 @@ class InvitationV1Spec extends ZSuite {
                                 |  "requests~attach": [
                                 |    {
                                 |      "@id": "request-0",
-                                |      "mime-type": "application/json",
                                 |      "data": {"base64" : "$payloadBase64"}
                                 |    }
                                 |  ],
@@ -53,7 +52,8 @@ class InvitationV1Spec extends ZSuite {
     val accepts = Seq("didcomm/aip2;env=rfc587", "didcomm/aip2;env=rfc19")
     val handshakeProtocols = Seq("https://didcomm.org/didexchange/1.0", "https://didcomm.org/connections/1.0")
 
-    val attachmentDescriptor = AttachmentDescriptor.buildAttachment(id = "request-0", payload = payload)
+    val attachmentDescriptor =
+      AttachmentDescriptor.buildBase64Attachment(id = "request-0", payload = payload.getBytes())
 
     val invitation = Invitation(
       `@id` = "f3375429-b116-4224-b55f-563d7ef461f1",