+++ pollux/lib/core/src/main/scala/io/iohk/atala/pollux/core/model/error/CredentialRepositoryError.scala
@@ -1,0 +1,7 @@
+package io.iohk.atala.pollux.core.model.error
+
+sealed trait CredentialRepositoryError extends Throwable
+
+object CredentialRepositoryError {
+  final case class UniqueConstraintViolation(message: String) extends CredentialRepositoryError
+}