+++ prism-agent/service/wallet-api/src/main/scala/io/iohk/atala/agent/walletapi/sql/JdbcDIDNonSecretStorage.scala
@@ -3,12 +3,13 @@ package io.iohk.atala.agent.walletapi.sql
 import doobie.*
 import doobie.implicits.*
 import doobie.postgres.implicits.*
-
-import io.iohk.atala.agent.walletapi.model.ManagedDIDState
+import io.iohk.atala.agent.walletapi.model.{DIDUpdateLineage, ManagedDIDState}
 import io.iohk.atala.agent.walletapi.storage.DIDNonSecretStorage
-import io.iohk.atala.castor.core.model.did.PrismDID
+import io.iohk.atala.castor.core.model.did.{PrismDID, ScheduledDIDOperationStatus}
 import zio.*
 import zio.interop.catz.*
+
+import java.time.Instant
 
 class JdbcDIDNonSecretStorage(xa: Transactor[Task]) extends DIDNonSecretStorage {
 
@@ -19,8 +20,10 @@ class JdbcDIDNonSecretStorage(xa: Transactor[Task]) extends DIDNonSecretStorage 
         |   did,
         |   publication_status,
         |   atala_operation_content,
-        |   publish_operation_id
-        | FROM public.did_publication_state
+        |   publish_operation_id,
+        |   created_at,
+        |   updated_at
+        | FROM public.prism_did_wallet_state
         | WHERE did = $did
         """.stripMargin
         .query[DIDPublicationStateRow]
@@ -32,27 +35,35 @@ class JdbcDIDNonSecretStorage(xa: Transactor[Task]) extends DIDNonSecretStorage 
   }
 
   override def setManagedDIDState(did: PrismDID, state: ManagedDIDState): Task[Unit] = {
-    val row = DIDPublicationStateRow.from(did, state)
-    val cxnIO = sql"""
-        | INSERT INTO public.did_publication_state(
+    val cxnIO = (row: DIDPublicationStateRow) => sql"""
+        | INSERT INTO public.prism_did_wallet_state(
         |   did,
         |   publication_status,
         |   atala_operation_content,
-        |   publish_operation_id
+        |   publish_operation_id,
+        |   created_at,
+        |   updated_at
         | )
         | VALUES (
         |   ${row.did},
         |   ${row.publicationStatus},
         |   ${row.atalaOperationContent},
-        |   ${row.publishOperationId}
+        |   ${row.publishOperationId},
+        |   ${row.createdAt},
+        |   ${row.updatedAt}
         | )
         | ON CONFLICT (did) DO UPDATE SET
         |   publication_status = EXCLUDED.publication_status,
         |   atala_operation_content = EXCLUDED.atala_operation_content,
-        |   publish_operation_id = EXCLUDED.publish_operation_id
+        |   publish_operation_id = EXCLUDED.publish_operation_id,
+        |   updated_at = EXCLUDED.updated_at
         """.stripMargin.update
 
-    cxnIO.run.transact(xa).unit
+    for {
+      now <- Clock.instant
+      row = DIDPublicationStateRow.from(did, state, now)
+      _ <- cxnIO(row).run.transact(xa)
+    } yield ()
   }
 
   override def listManagedDID: Task[Map[PrismDID, ManagedDIDState]] = {
@@ -62,8 +73,10 @@ class JdbcDIDNonSecretStorage(xa: Transactor[Task]) extends DIDNonSecretStorage 
            |   did,
            |   publication_status,
            |   atala_operation_content,
-           |   publish_operation_id
-           | FROM public.did_publication_state
+           |   publish_operation_id,
+           |   created_at,
+           |   updated_at
+           | FROM public.prism_did_wallet_state
            """.stripMargin
         .query[DIDPublicationStateRow]
         .to[List]
@@ -73,6 +86,72 @@ class JdbcDIDNonSecretStorage(xa: Transactor[Task]) extends DIDNonSecretStorage 
       .map(_.map(row => row.toDomain.map(row.did -> _)))
       .flatMap(ls => ZIO.foreach(ls)(ZIO.fromTry[(PrismDID, ManagedDIDState)](_)))
       .map(_.toMap)
+  }
+
+  override def insertDIDUpdateLineage(did: PrismDID, updateLineage: DIDUpdateLineage): Task[Unit] = {
+    val cxnIO =
+      sql"""
+           | INSERT INTO public.prism_did_update_lineage(
+           |   did,
+           |   operation_hash,
+           |   previous_operation_hash,
+           |   status,
+           |   operation_id,
+           |   created_at,
+           |   updated_at
+           | )
+           | VALUES (
+           |   $did,
+           |   ${updateLineage.operationHash},
+           |   ${updateLineage.previousOperationHash},
+           |   ${updateLineage.status},
+           |   ${updateLineage.operationId},
+           |   ${updateLineage.createdAt},
+           |   ${updateLineage.updatedAt}
+           | )
+           """.stripMargin.update
+
+    cxnIO.run.transact(xa).unit
+  }
+
+  override def listUpdateLineage(
+      did: Option[PrismDID],
+      status: Option[ScheduledDIDOperationStatus]
+  ): Task[Seq[DIDUpdateLineage]] = {
+    val didFilter = did.map(d => fr"did = $d")
+    val statusFilter = status.map(s => fr"status = $s")
+    val whereFr = Fragments.whereAndOpt(didFilter, statusFilter)
+    val baseFr =
+      sql"""
+           | SELECT
+           |   operation_id,
+           |   operation_hash,
+           |   previous_operation_hash,
+           |   status,
+           |   created_at,
+           |   updated_at
+           | FROM public.prism_did_update_lineage
+      """.stripMargin
+    val cxnIO = (baseFr ++ whereFr)
+      .query[DIDUpdateLineage]
+      .to[List]
+
+    cxnIO.transact(xa)
+  }
+
+  override def setDIDUpdateLineageStatus(
+      operationId: Array[Byte],
+      status: ScheduledDIDOperationStatus
+  ): Task[Unit] = {
+    val cxnIO = (now: Instant) => sql"""
+            | UPDATE public.prism_did_update_lineage
+            | SET
+            |   status = $status,
+            |   updated_at = $now
+            | WHERE operation_id = $operationId
+        """.stripMargin.update
+
+    Clock.instant.flatMap(now => cxnIO(now).run.transact(xa)).unit
   }
 
 }