+++ tests/e2e-tests/src/test/kotlin/features/did/ManageDidSteps.kt
@@ -7,7 +7,6 @@ import common.Utils.toJsonPath
 import io.cucumber.java.en.Given
 import io.cucumber.java.en.Then
 import io.cucumber.java.en.When
-import net.serenitybdd.rest.SerenityRest
 import net.serenitybdd.rest.SerenityRest.lastResponse
 import net.serenitybdd.screenplay.Actor
 import net.serenitybdd.screenplay.rest.interactions.Get
@@ -16,7 +15,6 @@ import net.serenitybdd.screenplay.rest.questions.ResponseConsequence
 import org.apache.http.HttpStatus.SC_OK
 import org.assertj.core.api.Assertions
 import org.hamcrest.Matchers.*
-import java.util.UUID
 
 class ManageDidSteps {